backend/pyproject.toml
[project]
name = "clarimetis-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "sqlalchemy[asyncio]==2.0.35",
    "asyncpg==0.30.0",
    "pgvector==0.3.5",
    "google-cloud-aiplatform==1.70.0",
    "google-cloud-storage==2.18.0",
    "google-cloud-speech==2.27.0",
    "google-cloud-texttospeech==2.17.0",
    "vertexai==1.70.0",
    "stripe==11.1.0",
    "PyJWT[cryptography]==2.9.0",
    "httpx==0.27.2",
    "python-multipart==0.0.12",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "python-dotenv==1.0.1",
    "tenacity==9.0.0",
    "filetype==1.2.0",
    "slowapi==0.1.9",
    "cachetools==5.5.2",
    "numpy==2.2.4",
    "langfuse==2.58.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
]
